# Conflicts:
#	out/production/Ristorante-IngegneriaDelSoftware/it/unibs/ingsw/restaurant/Handler.class
#	out/production/Ristorante-IngegneriaDelSoftware/it/unibs/ingsw/users/reservations_agent/ReservationItemList.class
#	out/production/Ristorante-IngegneriaDelSoftware/it/unibs/ingsw/users/reservations_agent/SimpleReservation.class
#	out/production/Ristorante-IngegneriaDelSoftware/it/unibs/ingsw/users/reservations_agent/agenda.xml
#	out/production/Ristorante-IngegneriaDelSoftware/it/unibs/ingsw/users/reservations_agent/archivio_prenotazioni/prenotazioni_04-07-2023.xml
#	out/production/Ristorante-IngegneriaDelSoftware/it/unibs/ingsw/users/warehouse_worker/WarehouseWorker.class
#	out/production/Ristorante-IngegneriaDelSoftware/it/unibs/ingsw/users/warehouse_worker/WarehouseWorkerController.class
#	src/it/unibs/ingsw/users/reservations_agent/agenda.xml

+++ src/it/unibs/ingsw/mylib/utilities/AsciiArt.java
@@ -405,7 +405,7 @@ public class AsciiArt {
      *
      * @param restaurantWorkload il carico sostenibile del ristorante.
      */
-    public static void seeInfoWorkload(int caricoRaggiunto, double restaurantWorkload){
+    public static void seeInfoWorkload(double caricoRaggiunto, double restaurantWorkload){
         System.out.println(UsefulStrings.ACTUAL_WORKLOAD_MESSAGE+(caricoRaggiunto));
         System.out.println(UsefulStrings.ACTUAL_WORKLOAD_AVAILABLE_MESSAGE+Math.floor((restaurantWorkload - caricoRaggiunto) *100)/100+"\n");
     }
